init: Omy's ARFlac.pl/ARCue.pl C port
Mostly to do with https://hydrogenaudio.org/index.php/topic,53583.0.html
thread about ARCue.pl. This is my own port/rewrite to C of the ARFlac.pl
script which was in turn based off ARCue.pl. As such then credit must go
to the authors of ARCue.pl and ARFlac.pl, although the actual C code is
my own work.

Basically, this is a C library (and reference program which uses it)
which can be used to check a directory of .flac files against the
AccurateRip database. See the thread linked above for discussion of it's
limitations.

https://hydrogenaudio.org/index.php/topic,61574.0.html